whisper-app/requirements.txt
openai-whisper
torch
pyaudio
numpy